Add Emacs EUnit utility feature: run current test function.
This determines the current test function under point (cursor), whether the active buffer
contains a test module (foo_tests) or the module itself (foo), then compiles the source
module (and test module if present) and runs that test function.

Since this added a third testing feature, there was some natural duplication that needed
to be factored out.  The notion of 'compile and run tests' was decoupled from exactly
which tests should be run.  The definitions of default keybindings were converted to a list
for clarity and conciseness.