Don't add newline after arrow on -callback lines
In the Emacs Erlang mode, typing "->" will usually automatically add a
newline, which is usually what you want when writing a function, but
not when writing a type spec.  Therefore, there is a check for not
adding the newline when the current line starts with -spec or -type.

This change adds -callback to that check, since it is usually written
the same way as type specs.