Fix line-initial commas' indentation in type specs
Inside type specs, line-initial commas shall align with open curly
braces. This also applies to record definitions with type specs for
the fields.

For example:

    -type foo() :: { a :: integer()
                   , b :: integer()
                   }.

For type specs for functions, the commas shall align with the opening
parenthese. So do a closing parenthese when put on a new line.

For example:

    -spec foo( X :: integer()
             , Y :: integer()
             ) -> integer().

This does not affect coding styles that don't put commas at the
beginning of lines.