Add an option erlang-icr-indent
The option makes it possible to customize the indention of
if/case/recieve patterns.

Before the change, the indentation of 'if' patterns is 3 steps and
'case' patterns is 4 steps.  It cannot be changed alone.  Some people
prefer other styles, for example, zero indentation.

case a of
true ->
   do_something();
false ->
   do_something_else()
end.

This patch just adds an option `erlang-icr-indent'. Its default value is
`nil' which means keeping the legacy behavior.  When non-nil, the
pattens of if/case/receive are indented according to
`erlang-icr-indent'.  "(setq erlang-icr-indent 0)" will keep
if/case/receive at the same column of the sub-clauses.

Change-Id: I10c32e42dbf69cbe187f55223b9aa7824853e493