Emacs: Guess type of tag definition in Emacs 24
This concerns tag finding commands in Emacs 24.  If the word at point
is preceded by a hash suggest as default a jump to a record
definition.  If it is a question mark suggest a macro.  Otherwise
assume it is a function.  If it is not a qualified call assume it is a
local function.