Emacs erlang-mode: Improve erlang-man-function's search
The Emacs lisp function erlang-man-function opens the manual page for
a module and searches within the manual page for the documentation of
the specified function. Prior to this commit, the search often failed
to find the text for the function even when the correct man page was
opened. This commit fixes this issue.