Emacs erlang-mode: Fix bug in erlang-man-function and add menu item
Prior to this commit, the Emacs lisp function erlang-man-function
would only go to the correct spot of the man page when a buffer for
the man page existed before the function call. This commit fixes this
issue. The commit also adds a menu item for the function
erlang-man-function-no-prompt (erlang-man-function-no-prompt opens the
documentation for a function without prompting the user).