Emacs erlang-mode: make show doc for function more robust
This commit makes the functionality for opening the documentation for
a function more robust than previously and should work even if the API
of the Emacs man package changes, given that the name of the man
buffer for a module contains the module name and the word man. The
previous fix broke the open documentation for a function functionality
in Emacs 24.3. The code in this commit has been tested to work in
Emacs 24.3, Emacs 25 and Emacs 26.